Update scope of external package imports
In some cases, having these packages imported on the global level
caused errors durring the installation process. This is because
some ChatterBot classes are imported in __init__.py files and
the main __init__.py in the chatterbot modue is imported by
setup.py. As a result, the external packages are imported before
they get installed.